Remove user infrastructure directory
