+++ api/src/infrastructure/services/background_task_queue.py
@@ -6,7 +6,6 @@ class BackgroundTaskQueue(ABC):
     def add_task(self, func, *args, **kwargs):
         pass
 
-# TODO: separate this out into a separate file when needed
 class FastAPIBackgroundTaskQueue(BackgroundTaskQueue):
     def __init__(self, background_tasks: BackgroundTasks):
         self.background_tasks = background_tasks
