+++ api/src/users/api/dependencies.py
@@ -3,9 +3,9 @@ from fastapi import Depends
 from src.infrastructure.db.database import get_session
 from src.users.application.use_cases.login import Login
 from src.users.application.use_cases.register_user import RegisterUser
-from src.users.infrastructure.repositories.user_repository import UserRepository
-from src.users.infrastructure.utilities.bcrypt_hasher import BCryptHasher
-from src.users.infrastructure.utilities.jwt_token_service import JWTTokenService
+from src.users.application.repositories.user_repository import UserRepository
+from src.users.application.utilities.hasher import BCryptHasher
+from src.users.application.utilities.token_service import JWTTokenService
 
 
 def get_user_repository(session=Depends(get_session)):
